requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
nba_api==1.2.1
